+ Moved to static library build output + Tried std::expected for error handling of shader compilation + Create graphics context abstraction class with OpenGL implementation

+++ Mist/src/Mist/ImGui/ImGuiLayer.cpp
@@ -1,7 +1,7 @@
 #include "mistpch.h"
 #include "ImGuiLayer.h"
 
-#include "Application.h"
+#include "Mist/Application.h"
 
 #include "backends/imgui_impl_glfw.h"
 #include "backends/imgui_impl_opengl3.h"
@@ -80,7 +80,7 @@ void ImGuiLayer::Begin() {
 void ImGuiLayer::End() {
     ImGuiIO& io = ImGui::GetIO();
     Application& app = Application::Get();
-    io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+    io.DisplaySize = ImVec2((float) app.GetWindow().GetWidth(), (float) app.GetWindow().GetHeight());
 
     ImGui::Render();
     ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
